[macros]: WIP - reexport proc macros to work around reexports
