+++ proc-macros/README.md
@@ -1,0 +1,10 @@
+# commonware-macros
+
+[![Crates.io](https://img.shields.io/crates/v/commonware-macros.svg)](https://crates.io/crates/commonware-macros)
+[![Docs.rs](https://docs.rs/commonware-macros/badge.svg)](https://docs.rs/commonware-macros)
+
+Augment the development of primitives with procedural macros.
+
+## Status 
+
+`commonware-macros` is **ALPHA** software and is not yet recommended for production use. Developers should expect breaking changes and occasional instability.
